tests/Node-appendChild.html

<!DOCTYPE html>
<title>Node.appendChild</title>
<script src="testharness.js"></script>
<div id="log"></div>
<iframe src=about:blank></iframe>
<script>
function testHRE(node) {
  test(function() {
    assert_throws("HIERARCHY_REQUEST_ERR", function() { node.appendChild(document.createTextNode("fail")) })
    assert_throws(new TypeError(), function() { node.appendChild(null) })
  })
}
test(function() {
  var frameDoc = frames[0].document
  var s = frameDoc.createElement("a")
  assert_equals(s.ownerDocument, frameDoc)
  document.body.appendChild(s)
  assert_equals(s.ownerDocument, document)

  assert_throws("HIERARCHY_REQUEST_ERR", function() { document.body.appendChild(frameDoc) })

  s = frameDoc.createElement("b")
  frameDoc.body.appendChild(s)
  assert_equals(s.ownerDocument, frameDoc)
  document.body.appendChild(s)
  assert_equals(s.ownerDocument, document)

  assert_throws(new TypeError(), function() { document.body.appendChild(null) })

  assert_throws(new TypeError(), function() { document.body.appendChild({'a':'b'}) })

  document.body.setAttribute("abc", "pass")
  var attr = document.body.attributes[0]
  testHRE(attr);
  assert_equals(attr.value, "pass")
  assert_equals(attr.childNodes.length, 1)
  assert_equals(attr.childNodes[0].nodeValue, "pass")

  testHRE(document.createTextNode("Foo"));
})
</script>
